+++ internal/helpers/bucket_helpers.go
@@ -1,8 +1,17 @@
 package helpers
 
-import "fmt"
-import "regexp"
-import "strings"
+import (
+	"bufio"
+	"context"
+	"fmt"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/s3"
+	"log"
+	"os"
+	"regexp"
+	"strings"
+	"strconv"
+)
 
 const (
 	IsBucketRequestedSuffix = "-bucket-requested"
@@ -53,6 +62,61 @@ func ValidateBucketName(name string) bool {
 		disallowed = regexp.MustCompile(fmt.Sprintf("[^%s]+", whitelist))
 	)
 	return !disallowed.MatchString(name)
+}
+
+func GetBucketRequestLimit(ctx context.Context) int {
+	var maxBucketsEnv = os.Getenv("S3_MAX_BUCKETS_PER_REQUEST")
+	var maxBuckets, err = strconv.Atoi(maxBucketsEnv)
+
+	if err != nil {
+		log.Fatalf("Unable to read max buckets per request environment variable due to : %v", err)
+	}
+	return maxBuckets
+}
+
+// getBuckets retrieves a list of valid bucket names from an S3 object, validates them, and enforces a maximum limit.
+func GetBuckets(ctx context.Context, client *s3.Client, bucket string, key string) []string {
+	var buckets []string
+
+	resp, err := client.GetObject(ctx, &s3.GetObjectInput{
+		Bucket: aws.String(bucket),
+		Key:    aws.String(key),
+	})
+
+	if err != nil {
+		log.Fatalf("failed to get object: %s from %s due to %s", key, bucket, err)
+		return nil
+	}
+	defer resp.Body.Close()
+
+	scanner := bufio.NewScanner(resp.Body)
+
+	for scanner.Scan() {
+		line := scanner.Text()
+		log.Printf("Reading bucket name: %s", line)
+		if ValidateBucketName(line) {
+			buckets = append(buckets, line)
+		} else {
+			log.Fatalf("invalid bucket name requested: %s", line)
+			return nil
+		}
+	}
+
+	if err := scanner.Err(); err != nil {
+		log.Fatalf("Error reading response: %v", err)
+		return nil
+	}
+
+	// TODO: do we want to error like this or ignore extras (i.e. don't append additional buckets above)?
+	var maxBuckets = GetBucketRequestLimit(ctx)
+	bucketsRequested := len(buckets)
+	if bucketsRequested >= maxBuckets {
+		log.Fatalf("Exceeded maximum allowed buckets per request [%s] with [%s]",
+			maxBuckets, bucketsRequested)
+		return nil
+	}
+
+	return buckets
 }
